Fix buddy migrations: change to unsignedInteger to match users table
Changed sender_id, receiver_id, user_id, and buddy_id from unsignedBigInteger
to unsignedInteger to match the users table id column type (increments).

This fixes the foreign key constraint error during migration.

+++ database/migrations/2025_10_24_110000_create_buddy_requests_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('buddy_requests', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('sender_id');
-            $table->unsignedBigInteger('receiver_id');
+            $table->unsignedInteger('sender_id');
+            $table->unsignedInteger('receiver_id');
             $table->text('message')->nullable();
             $table->enum('status', ['pending', 'accepted', 'rejected'])->default('pending');
             $table->timestamps();
